制作了角色死亡状态和死亡弹窗提示，以及简单的复活debug功能
- Added PlayerUIPopUpManager script to handle "YOU DIED" pop-up functionality, including fade-in and fade-out animations.
- Updated layout files (CurrentMaximizeLayout.dwlt and default-6000.dwlt) with new control IDs and adjusted properties for UI elements.
- Modified various parameters in layout files to improve UI responsiveness and visual consistency.

+++ Assets/Scripts/Save & Load/CharacterSaveData.cs
@@ -23,7 +23,7 @@ namespace SG
         public float zPos; //角色在世界中的Z坐标
 
         [Header("Resources")]
-        public float currentHealth; //当前生命值
+        public int currentHealth; //当前生命值
         public float currentStamina; //当前耐力值
 
         [Header("Character Stats")]
